requirements.txt
Django==4.2.4
djangorestframework==3.14.0
Markdown==3.4.4

